refactor: add barrel file for billing enums, update imports to use it, refs AC-1202
