+++ apps/web/src/app/admin-console/settings/sponsored-families.component.ts
@@ -9,7 +9,7 @@ import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUti
 import { StateService } from "@bitwarden/common/abstractions/state.service";
 import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { Organization } from "@bitwarden/common/admin-console/models/domain/organization";
-import { PlanSponsorshipType } from "@bitwarden/common/billing/enums/plan-sponsorship-type";
+import { PlanSponsorshipType } from "@bitwarden/common/billing/enums/";
 import { SyncService } from "@bitwarden/common/vault/abstractions/sync/sync.service.abstraction";
 
 interface RequestSponsorshipForm {
