+++ apps/web/src/app/billing/settings/organization-plans.component.ts
@@ -25,8 +25,7 @@ import { OrganizationCreateRequest } from "@bitwarden/common/admin-console/model
 import { OrganizationKeysRequest } from "@bitwarden/common/admin-console/models/request/organization-keys.request";
 import { OrganizationUpgradeRequest } from "@bitwarden/common/admin-console/models/request/organization-upgrade.request";
 import { ProviderOrganizationCreateRequest } from "@bitwarden/common/admin-console/models/request/provider/provider-organization-create.request";
-import { PaymentMethodType } from "@bitwarden/common/billing/enums/payment-method-type";
-import { PlanType } from "@bitwarden/common/billing/enums/plan-type";
+import { PaymentMethodType , PlanType } from "@bitwarden/common/billing/enums";
 import { PlanResponse } from "@bitwarden/common/billing/models/response/plan.response";
 import { ProductType } from "@bitwarden/common/enums";
 import { EncString } from "@bitwarden/common/models/domain/enc-string";
