+++ libs/common/src/admin-console/models/request/organization/organization-sponsorship-create.request.ts
@@ -1,4 +1,4 @@
-import { PlanSponsorshipType } from "../../../../billing/enums/plan-sponsorship-type";
+import { PlanSponsorshipType } from "../../../../billing/enums";
 
 export class OrganizationSponsorshipCreateRequest {
   sponsoredEmail: string;
